+++ tests/test_framework.c
@@ -6,7 +6,7 @@
 /*   By: guillsan <guillsan@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 10:07:45 by guillsan          #+#    #+#             */
-/*   Updated: 2025/10/14 19:41:49 by guillsan         ###   ########.fr       */
+/*   Updated: 2025/10/14 22:29:22 by guillsan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,14 +24,14 @@ float			g_flush_seconds = 0.02;
 
 typedef struct s_sub
 {
-    int idx;
-    int ok;
+    size_t idx;
+    size_t ok;
     char *detail;
 } t_sub;
 
 static const char *g_group_name = NULL;
 static t_sub g_subs[MAX_SUBTESTS];
-static int g_sub_count = 0;
+static size_t g_sub_count = 0;
 
 void group_start(const char *name)
 {
@@ -42,7 +42,7 @@ void group_start(const char *name)
 void group_subtest(int ok, const char *fmt, ...)
 {
     if (g_sub_count >= MAX_SUBTESTS) return;
-    int i = g_sub_count;
+    size_t i = g_sub_count;
     g_subs[i].idx = i + 1;
     g_subs[i].ok = ok;
     g_subs[i].detail = NULL;
@@ -60,16 +60,16 @@ void group_subtest(int ok, const char *fmt, ...)
 }
 
 /* finish group: print grouped line then detailed failures, return failures count */
-int group_finish()
+size_t group_finish()
 {
     /* Print compact group header + statuses */
 	printf("\n%s%-12s%s:", CLR_CYAN, g_group_name ? g_group_name : "(unknown)", CLR_RESET);
-    for (int i = 0; i < g_sub_count; ++i)
+    for (size_t i = 0; i < g_sub_count; ++i)
     {
 		if (g_subs[i].ok)
-            printf(" %s%d.OK%s%s ", CLR_GREEN, g_subs[i].idx, SYMBOL_OK, CLR_RESET);
+            printf(" %s%ld.OK%s%s ", CLR_GREEN, g_subs[i].idx, SYMBOL_OK, CLR_RESET);
         else
-            printf(" %s%d.KO%s%s ", CLR_RED, g_subs[i].idx, SYMBOL_KO, CLR_RESET);
+            printf(" %s%ld.KO%s%s ", CLR_RED, g_subs[i].idx, SYMBOL_KO, CLR_RESET);
         
         if (g_b_flush)
         {
@@ -79,15 +79,15 @@ int group_finish()
     }
     
     /* Print details for failures (one-by-one) */
-    int failures = 0;
-    int has_errors = 0;
-    for (int i = 0; i < g_sub_count; ++i)
+    size_t failures = 0;
+    size_t has_errors = 0;
+    for (size_t i = 0; i < g_sub_count; ++i)
     {
         if (!g_subs[i].ok)
         {
             failures++;
 			has_errors = 1;
-			printf(INDENT "\n%s%s (%d):%s %s", CLR_YELLOW, g_group_name, g_subs[i].idx, CLR_RESET,
+			printf(INDENT "\n%s%s (%ld):%s %s", CLR_YELLOW, g_group_name, g_subs[i].idx, CLR_RESET,
                    g_subs[i].detail ? g_subs[i].detail : "failed");
             free(g_subs[i].detail);
             g_subs[i].detail = NULL;
