+++ tests/test_framework.h
@@ -6,12 +6,13 @@
 /*   By: guillsan <guillsan@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 10:06:33 by guillsan          #+#    #+#             */
-/*   Updated: 2025/10/14 20:17:48 by guillsan         ###   ########.fr       */
+/*   Updated: 2025/10/14 22:26:47 by guillsan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef TEST_FRAMEWORK_H
 # define TEST_FRAMEWORK_H
+#include <stddef.h>
 
 /* ANSI colors & symbols */
 # define CLR_RESET 	"\x1b[0m"
@@ -29,55 +30,55 @@
 /* group API used by test files */
 void group_start(const char *name);
 void group_subtest(int ok, const char *fmt, ...);
-int  group_finish();
+size_t  group_finish();
 
 /* convenience macros */
 # define SUBTEST_OK()            (group_subtest(1, NULL))
 # define SUBTEST_FAILF(...)      (group_subtest(0, __VA_ARGS__))
 
-int group_isalpha(void);
-int group_isdigit(void);
-int group_isalnum(void);
-int group_isascii(void);
-int group_isprint(void);
-int group_strlen(void);
-int group_memset(void);
-int group_bzero(void);
-int group_memcpy(void);
-int group_memmove(void);
-int group_strlcpy(void);
-int group_strlcat(void);
-int group_toupper(void);
-int group_tolower(void);
-int group_strchr(void);
-int group_strrchr(void);
-int group_strncmp(void);
-int group_memchr(void);
-int group_memcmp(void);
-int group_strnstr(void);
-int group_atoi(void);
-int group_calloc(void);
-int group_strdup(void);
-int group_substr(void);
-int group_strjoin(void);
-int group_strtrim(void);
-int group_split(void);
-int group_itoa(void);
-int group_strmapi(void);
-int group_striteri(void);
-int group_putchar_fd(void);
-int group_putstr_fd(void);
-int group_putendl_fd(void);
-int group_putnbr_fd(void);
+size_t group_isalpha(void);
+size_t group_isdigit(void);
+size_t group_isalnum(void);
+size_t group_isascii(void);
+size_t group_isprint(void);
+size_t group_strlen(void);
+size_t group_memset(void);
+size_t group_bzero(void);
+size_t group_memcpy(void);
+size_t group_memmove(void);
+size_t group_strlcpy(void);
+size_t group_strlcat(void);
+size_t group_toupper(void);
+size_t group_tolower(void);
+size_t group_strchr(void);
+size_t group_strrchr(void);
+size_t group_strncmp(void);
+size_t group_memchr(void);
+size_t group_memcmp(void);
+size_t group_strnstr(void);
+size_t group_atoi(void);
+size_t group_calloc(void);
+size_t group_strdup(void);
+size_t group_substr(void);
+size_t group_strjoin(void);
+size_t group_strtrim(void);
+size_t group_split(void);
+size_t group_itoa(void);
+size_t group_strmapi(void);
+size_t group_striteri(void);
+size_t group_putchar_fd(void);
+size_t group_putstr_fd(void);
+size_t group_putendl_fd(void);
+size_t group_putnbr_fd(void);
 
-int group_lstnew(void);
-int group_lstadd_front(void);
-int group_lstsize(void);
-int group_lstlast(void);
-int group_lstadd_back(void);
-int group_lstdelone(void);
-int group_lstclear(void);
-int group_lstiter(void);
-int group_lstmap(void);
+size_t group_lstnew(void);
+size_t group_lstadd_front(void);
+size_t group_lstsize(void);
+size_t group_lstlast(void);
+size_t group_lstadd_back(void);
+size_t group_lstdelone(void);
+size_t group_lstclear(void);
+size_t group_lstiter(void);
+size_t group_lstmap(void);
 
 #endif /* TEST_FRAMEWORK_H */
