+++ tests/test_isalnum.c
@@ -6,7 +6,7 @@
 /*   By: guillsan <guillsan@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 11:45:23 by guillsan          #+#    #+#             */
-/*   Updated: 2025/10/14 20:30:38 by guillsan         ###   ########.fr       */
+/*   Updated: 2025/10/14 22:19:28 by guillsan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include "../../libft.h"
 #define SIGN_OK(e,g) ((exp == 0 && got != 0) || (exp != 0 && got == 0))
 
-int group_isalnum(void)
+size_t group_isalnum(void)
 {
     group_start("isalnum");
