+++ tests/test_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: guillsan <guillsan@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 10:16:07 by guillsan          #+#    #+#             */
-/*   Updated: 2025/10/14 12:47:00 by guillsan         ###   ########.fr       */
+/*   Updated: 2025/10/14 22:22:43 by guillsan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 #include <string.h>
 #include "../../libft.h"
 
-int group_memcpy(void)
+size_t group_memcpy(void)
 {
     group_start("memcpy");
